+++ commitlint.config.ts
@@ -1,0 +1,3 @@
+export default {
+  extends: '@vfourny/node-toolkit/commitlint',
+}