+++ eslint.config.js
@@ -1,0 +1,4 @@
+import typescriptEslint from 'typescript-eslint'
+import nodeToolkitEslint from '@vfourny/node-toolkit/eslint'
+
+export default typescriptEslint.config(...nodeToolkitEslint)