+++ src/components/UserSession.component.vue
@@ -1,17 +1,17 @@
 <template>
   <div>
     <div v-if="isAuthenticated" class="text-center">
-      <p class="text-lg">{{ t("userSession.welcome") }}, {{ user?.name }}!</p>
+      <p class="text-lg">{{ t('userSession.welcome') }}, {{ user?.name }}!</p>
       <button
         @click="logout"
         class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded mt-4"
       >
-        {{ t("userSession.logout") }}
+        {{ t('userSession.logout') }}
       </button>
     </div>
     <div v-else>
       <p class="text-lg mb-4">
-        {{ t("userSession.askLogin") }}
+        {{ t('userSession.askLogin') }}
       </p>
       <input
         v-model="email"
@@ -27,21 +27,21 @@
         @click="login"
         class="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
       >
-        {{ t("userSession.login") }}
+        {{ t('userSession.login') }}
       </button>
     </div>
   </div>
 </template>
 <script setup lang="ts">
-import { ref, computed } from "vue"
-import { useSessionStore } from "../stores"
-import { useI18n } from "vue-i18n"
+import { ref, computed } from 'vue'
+import { useSessionStore } from '../stores'
+import { useI18n } from 'vue-i18n'
 
 const { t } = useI18n()
 
 const sessionStore = useSessionStore()
-const email = ref("")
-const name = ref("")
+const email = ref('')
+const name = ref('')
 
 const login = () => {
   const user = {