+++ src/router/index.ts
@@ -1,8 +1,8 @@
-import { createMemoryHistory, createRouter } from "vue-router"
+import { createMemoryHistory, createRouter } from 'vue-router'
 
-import HomeSreenPage from "../pages/HomeScreen.page.vue"
+import HomeSreenPage from '../pages/HomeScreen.page.vue'
 
-const routes = [{ path: "/", component: HomeSreenPage }]
+const routes = [{ path: '/', component: HomeSreenPage }]
 
 const router = createRouter({
   history: createMemoryHistory(),