+++ src/stores/index.ts
@@ -1,3 +1,3 @@
-import useSessionStore from "./session.store"
+import useSessionStore from './session.store'
 
 export { useSessionStore }