+++ docs/CHANGELOG.md
@@ -1,7 +1,6 @@
 # 1.0.0 (2025-02-04)
 
-
 ### Bug Fixes
 
-* add nodetoolkit config ([#3](https://github.com/vfourny/template-vue/issues/3)) ([caf7dac](https://github.com/vfourny/template-vue/commit/caf7daca69326065e444e4a1e7df1266420468ca))
-* init project ([0333d8f](https://github.com/vfourny/template-vue/commit/0333d8f9f95b49a0b2fbb85de3727e5417ada5bb))
+- add nodetoolkit config ([#3](https://github.com/vfourny/template-vue/issues/3)) ([caf7dac](https://github.com/vfourny/template-vue/commit/caf7daca69326065e444e4a1e7df1266420468ca))
+- init project ([0333d8f](https://github.com/vfourny/template-vue/commit/0333d8f9f95b49a0b2fbb85de3727e5417ada5bb))